+++ CHANGELOG.md
@@ -9,9 +9,17 @@ Packaged by Steven Czerwinski <czerwin@scalyr.com> on Jun 18, 2015 13:00 -0700
 
 Note, this release is still under development.  Release date and details are TBD.
 
+New features:
+
+* Allow any monitor's metric log's rate limiter to be modified via configuration, along with flush aggregation.
+* Provide ``--upgrade-without-ui`` on ``scalyr-agent-2-config`` to allow upgrading the agent without UI (Windows).
+* Write user-friendly error message to event log when a major configuration problem is seen during start up (Windows)
+
 Bug fixes:
 
-* Do not fail Scalyr Agent service start if configuration file registry is not set (Windows)
+* Do not fail Scalyr Agent service start if configuration file registry is not set (Windows).
+* Remove rate limiter on metric log for graphite monitor and add in flush aggregation.
+* Fix that prevented graphite monitor thread from starting when accepting both text and pickle protocols
 
 ## 2.0.10 "Jumpy Jaguar" - June  9, 2015